rename ppnlp_kernel
