+++ ops/setup.py
@@ -16,6 +16,7 @@ import os
 import shutil
 import sys
 import textwrap
+import warnings
 from pathlib import Path
 
 from setuptools import find_packages, setup
@@ -88,7 +89,7 @@ def write_custom_op_api_py(libname, filename):
 
 if len(sys.argv) > 0:
     # generate lib files
-    lib_path = Path("src/paddlenlp_gpu_ops/cuda/lib")
+    lib_path = Path("src/ppnlp_kernel/cuda/lib")
     if lib_path.exists():
         shutil.rmtree(lib_path)
     lib_path.mkdir(exist_ok=True)
@@ -107,23 +108,23 @@ if len(sys.argv) > 0:
         has_built = True
 
     if not has_built:
-        raise RuntimeError("No cuda lib found. Please build cuda lib first. See details in csrc/README.md")
+        warnings.warn("No cuda lib found. Please build cuda lib first. See details in csrc/README.md.")
 
 # NEW ADDED END
 setup(
-    name="paddlenlp_gpu_ops",
+    name="ppnlp_kernel",
     version="0.1.0",  # expected format is one of x.y.z.dev0, or x.y.z.rc1 or x.y.z (no to dashes, yes to dots)
     description="PaddleNLP GPU OPS cuda & triton.",
     long_description=open("README.md", "r", encoding="utf-8").read(),
     long_description_content_type="text/markdown",
-    keywords="paddlenlp gpu ops contain cuda & triton",
+    keywords="paddlenlp kernel contain cuda & triton",
     license="Apache 2.0 License",
     author="PaddlePaddle",
     author_email="paddlenlp@baidu.com",
-    url="https://github.com/PaddlePaddle/paddlenlp/csrc/ops",
+    url="https://github.com/PaddlePaddle/paddlenlp/ops",
     package_dir={"": "src"},
     packages=find_packages("src"),
-    package_data={"paddlenlp_gpu_ops.cuda.lib": ["*.so", "*.dll", "*.dylib"]},
+    package_data={"ppnlp_kernel.cuda.lib": ["*.so", "*.dll", "*.dylib"]},
     include_package_data=True,
     python_requires=">=3.8.0",
     install_requires=read_requirements_file("requirements.txt"),
