+++ ops/utils/release.py
@@ -25,7 +25,7 @@ REPLACE_PATTERNS = {
     "setup": (re.compile(r'^(\s*)version\s*=\s*"[^"]+",', re.MULTILINE), r'\1version="VERSION",'),
 }
 REPLACE_FILES = {
-    "init": "src/paddlenlp_gpu_ops/__init__.py",
+    "init": "src/ppnlp_kernel/__init__.py",
     "setup": "setup.py",
 }
 README_FILE = "README.md"
